urbanroof-ddr-updated/requirements.txt
pymupdf==1.24.0
google-generativeai==0.7.2
streamlit==1.35.0
Pillow==10.3.0
fpdf2==2.7.9
reportlab